services/document-converter/requirements.txt
# RabbitMQ
pika>=1.3.0

# Azure services
azure-ai-formrecognizer>=3.3.0
openai>=1.0.0

# Document processing
python-docx>=1.1.0
pandas>=2.0.0
openpyxl>=3.1.0
pdf2image>=1.16.0
PyMuPDF>=1.23.0

# Object storage
minio>=7.2.0

# Utilities
charset-normalizer>=3.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-magic-bin>=0.4.14;platform_system=="Windows"

# Async support
asyncio>=3.4.3;python_version<"3.9"

# Logging
python-json-logger>=2.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
